Minor Changes.
Fixed bugs.

*The app is currently in process of re-building*
*So please DO NOT use the current version of the app*
