+++ src/Launcher.java
@@ -1,10 +1,11 @@
-import static ui.layout.Logos.logoInitializer;
-import static ui.layout.Stylization.drawFullTripleBorder;
-import static ui.layout.UiLayout.nameOfLogo;
+import static ui.layout.BorderWork.drawFullTripleBorder;
+import static ui.layout.LogoWork.logoInitializer;
+import static ui.layout.LogoWork.nameOfLogo;
 import static ui.pages.MenuPage.displayMainMenuUi;
 
 public class Launcher {
     public static void main(String[] args) {
+        //displaySetUpPage();
         drawFullTripleBorder();
         logoInitializer(nameOfLogo);
         drawFullTripleBorder();
