+++ src/commands_language_packages/EnglishCommands.java
@@ -1,7 +1,7 @@
 package commands_language_packages;
 
 import additional_packages.calculator.Calculator;
-import ui.layout.UiLayout;
+import ui.layout.AdditionalFunctions;
 import ui.pages.GeneralSettingsPage;
 import ui.pages.InfoPage;
 import ui.pages.TimePage;
@@ -16,13 +16,13 @@ public class EnglishCommands {
         commandMap.put("settings", GeneralSettingsPage::displayGeneralSettings);
         commandMap.put("set", GeneralSettingsPage::displayGeneralSettings);
 
-        commandMap.put("commands", UiLayout::displayListOfMenuCommands);
-        commandMap.put("com", UiLayout::displayListOfMenuCommands);
+        commandMap.put("commands", AdditionalFunctions::displayListOfMenuCommands);
+        commandMap.put("com", AdditionalFunctions::displayListOfMenuCommands);
 
         commandMap.put("time", TimePage::displayCurrentTime);
 
         commandMap.put("info", InfoPage::displayInfo);
 
-        commandMap.put("exit", UiLayout::exitProgram);
+        commandMap.put("exit", AdditionalFunctions::exitProgram);
     }
 }
