--- src/ui/layout/Borders.java
@@ -1,4 +1,0 @@
-package ui.layout;
-
-public class Borders {
-}
