+++ src/ui/layout/TextWork.java
@@ -1,0 +1,51 @@
+package ui.layout;
+
+import static ui.layout.BorderWork.*;
+import static ui.layout.ColorWork.*;
+
+public class TextWork {
+
+    public static void wrapText(String text, int width) {
+        for (int i = 0; i < text.length(); i += width) {
+            if (i == 0) {
+                int end = Math.min(i + width, text.length());
+                drawHorizontalBorder(numberOfSymbols);
+                System.out.print(BOLD + contentAlignment(text.length() + 2) + "·" + text.substring(i, end) + "·" + RESET);
+                if (end < text.length()) {
+                    System.out.println();
+                }
+            } else {
+                int end = Math.min(i + width, text.length());
+                System.out.print(BOLD + contentAlignment(text.length() + 2) + "·" + text.substring(i, end) + "·" + RESET);
+                if (end < text.length()) {
+                    System.out.println();
+                }
+            }
+        }
+        System.out.println();
+    }
+
+    //Show slow motion text
+    public static void displaySlowMotionText(int delay, int alignment, boolean isUnderlineActive, String mainText, String additionalText) {
+        String formattedText = contentAlignment(alignment) +
+                (isUnderlineActive ? UNDERLINE : "") +
+                BOLD + mainText + RESET + additionalText;
+
+        for (char ch : formattedText.toCharArray()) {
+            System.out.print(ch);
+            try {
+                Thread.sleep(delay);
+            } catch (InterruptedException ex) {
+                displayColorCommand("Error, try again", "red", (byte) 0);
+            }
+        }
+        System.out.print("");
+    }
+
+    //Aligns the content
+    public static String contentAlignment(int widthOfElement) {
+        int fullWidth = borderWidth + 2; //Where 2 is "+"
+        int oneSide = (fullWidth - widthOfElement) / 2;
+        return " ".repeat(Math.max(0, oneSide));
+    }
+}
