+++ src/ui/pages/GeneralSettingsPage.java
@@ -1,12 +1,14 @@
 package ui.pages;
 
-import ui.layout.UiLayout;
+import ui.layout.AdditionalFunctions;
 
 import java.util.HashMap;
 
-import static ui.layout.Stylization.*;
-import static ui.layout.UiLayout.displayTip;
-import static ui.layout.UiLayout.scanner;
+import static ui.layout.AdditionalFunctions.displayTip;
+import static ui.layout.AdditionalFunctions.scanner;
+import static ui.layout.BorderWork.*;
+import static ui.layout.ColorWork.displayColorCommand;
+import static ui.layout.TextWork.*;
 import static ui.pages.MenuPage.nameOfFunction;
 
 public class GeneralSettingsPage {
@@ -14,16 +16,16 @@ public class GeneralSettingsPage {
     public static void displayGeneralSettings() {
         HashMap<String, Runnable> listOfSettings = new HashMap<>();
 
-        listOfSettings.put("commands", UiLayout::displayListOfSettings);
-        listOfSettings.put("settings value", UiLayout::displayListOfMenuCommands);
-        listOfSettings.put("logo", LogoSwitcher::logoSwitcherUi);
-        listOfSettings.put("border", LogoSwitcher::logoSwitcherUi); //In progress
-        listOfSettings.put("delay", LogoSwitcher::logoSwitcherUi); //In progress
-        listOfSettings.put("exit", UiLayout::exitBlock);
+        listOfSettings.put("commands", AdditionalFunctions::displayListOfSettings);
+        listOfSettings.put("settings value", AdditionalFunctions::displayListOfMenuCommands);
+        listOfSettings.put("logo", LogoSwitcherPage::logoSwitcherUi);
+        listOfSettings.put("border", LogoSwitcherPage::logoSwitcherUi); //In progress
+        listOfSettings.put("delay", LogoSwitcherPage::logoSwitcherUi); //In progress
+        listOfSettings.put("exit", AdditionalFunctions::exitBlock);
 
         displaySlowMotionText(100, 8, false, "Settings\n", "");
 
-        drawFullTripleBorder();
+        drawHorizontalBorder(numberOfSymbols);
         displayTip("Enter 'commands'\n"
                 + contentAlignment(18) + "to show list of\n"
                 + contentAlignment(18) + "commands");
@@ -32,14 +34,14 @@ public class GeneralSettingsPage {
 
         nameOfFunction = scanner.nextLine().toLowerCase();
         wrapText(nameOfFunction, borderWidth - 2); //Wrapping the command from user
-        drawFullTripleBorder();
+        drawHorizontalBorder(numberOfSymbols);
 
         if (listOfSettings.containsKey(nameOfFunction)) {
             listOfSettings.get(nameOfFunction).run();
         } else {
-            drawFullTripleBorder();
+            drawHorizontalBorder(numberOfSymbols);
             displayColorCommand("Command not found", "red", (byte) 0);
-            drawFullTripleBorder();
+            drawHorizontalBorder(numberOfSymbols);
         }
     }
 }
