+++ src/ui/pages/InfoPage.java
@@ -1,17 +1,15 @@
 package ui.pages;
 
-import static ui.layout.Stylization.contentAlignment;
-import static ui.layout.Stylization.displayColorCommand;
+import static ui.layout.BorderWork.drawHorizontalBorder;
+import static ui.layout.BorderWork.numberOfSymbols;
+import static ui.layout.ColorWork.displayColorCommand;
+import static ui.layout.TextWork.contentAlignment;
 
 public class InfoPage {
     public static void displayInfo() {
-        //IN PROGRESS
-        displayVersion();
-    }
-
-    //Show version
-    public static void displayVersion() {
-        System.out.println(contentAlignment(18) + "Current version:");
-        displayColorCommand("0.4.8", "randomly", (byte) 0);
+        displayColorCommand(contentAlignment(18) + "Current version:", "white", (byte) 18);
+        displayColorCommand("0.4.9", "randomly", (byte) 18);
+        displayColorCommand("Author: Nick Zozulia", "white", (byte) 18);
+        drawHorizontalBorder(numberOfSymbols);
     }
 }
